bugfix: Fix recompilation after failed build
When a source file fails to compile, an object file of a previous valid state might exist.
If the user retries to build without modifying the file, the system could incorrectly treat the object file as up-to-date.
The introduced solution is to remove the object file upon failed compilation to guarantee a clean rebuild.