bugfix: Could not accept configuration names as argument if they have the same name as the command.
The command `build` could not accept a configuration called "build" and output an error about a duplicated flag.
The same issue occurred with the rest of the commands that accept a configuration name.

Fixed the issue and added some tests to make sure they do not reoccur in the future.